requirements.txt
aiogram
langchain-openai
openai
python-dotenv